kernelci.cli: add `AttributesCommand` class
Add a base class to use in different kci tools like
`kci user` and `kci node` to get attributes as command
line options.
This is to resolve pylint `duplicate-code` error from
`kernelci.api.user` and `kernelci.api.node` modules.
